refactor reshape

+++ src/operators/tensor/implementations/tensor_bool.cairo
@@ -60,7 +60,7 @@ impl BoolTensor of TensorTrait<bool> {
         unravel_index(index, *self.shape)
     }
 
-    fn reshape(self: @Tensor<bool>, target_shape: Span<usize>) -> Tensor<bool> {
+    fn reshape(self: @Tensor<bool>, target_shape: Span<i32>) -> Tensor<bool> {
         reshape(self, target_shape)
     }
